requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.23
protobuf==3.15.8
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
